functions/version1/faceanalyzerfn/requirements.txt
minio==7.1.1
numpy
Pillow==8.4.0
tensorflow
opencv-python-headless
requests